requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
charset-normalizer==3.3.2
pdfplumber==0.11.4
pypdf==4.3.1
faiss-cpu==1.8.0.post1
langchain==0.3.1
langchain-chroma==0.1.4
langchain-core==0.3.7
langchain-community==0.3.1
langchain-upstage==0.3.0
pydantic==2.9.2
streamlit==1.22.0
streamlit-chat==0.0.2.2
python-dotenv==1.0.1
langsmith